requirements.txt
cryptography==3.3.1
Django==2.2.6
djangorestframework==3.10.3
djangorestframework-simplejwt==4.6.0
django-cors-headers==3.6.0
Pillow==8.1.0
djoser==2.1.0
django-filter==2.4.0
django-autofixture==0.12.1
drf-yasg==1.20.0
django-grappelli==2.15.1
